refactor: App.tsx에서 UI 페이지로 분리
